NSObject: explicitly copy observer blocks
Block_Copy is required whenever you retain a block past the lifetime of its
scope, since the original block is stack based.

ARC handles this, but does so too late in this case -- we don't need a 
retained block until the addObject, so the __bridge cast ends up keeping
a pointer to a stack block.

By doing an explicit copy, we heap-allocate the block and get the expected
lifetime semantics -- the retained heap block is both in the array and 
used as context.